requirements.txt
# Core Dependencies
crewai>=0.15.0
langchain-community>=0.0.21
langchain-ollama>=0.1.3
numpy>=1.26.0

# LLM API Support
openai>=1.97.0
anthropic>=0.25.0

# Configuration and Environment
python-dotenv>=1.0.0
pydantic>=2.11.0
pydantic-settings>=2.0.0

# Development and Testing
pytest>=8.0.0
pytest-cov>=4.0.0
pytest-asyncio>=0.24.0
black>=24.0.0
flake8>=7.0.0
mypy>=1.8.0